[improvement] rename docs/src to docs/site (#3660)
