+++ docs/site/components/DemoList.vue
@@ -16,7 +16,7 @@
     </div>
 
     <h1 class="vant-title">
-      <img src="https://img.yzcdn.cn/vant/logo.png" >
+      <img src="https://img.yzcdn.cn/vant/logo.png">
       <span>Vant</span>
     </h1>
     <h2 class="vant-desc">{{ description }}</h2>
