+++ test/demo-test.ts
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import '../docs/src/demo-common';
+import '../docs/site/demo-common';
 import Locale from '../src/locale';
 import { mount, later } from './utils';
